+++ demo/boot/astro.config.ts
@@ -1,9 +1,9 @@
-import node from "@astrojs/node";
-import { defineConfig } from "astro/config";
-import boot from "@astroscope/boot";
+import node from '@astrojs/node';
+import boot from '@astroscope/boot';
+import { defineConfig } from 'astro/config';
 
 export default defineConfig({
-  output: "server",
-  adapter: node({ mode: "standalone" }),
+  output: 'server',
+  adapter: node({ mode: 'standalone' }),
   integrations: [boot()],
 });