+++ demo/proxy/astro.config.ts
@@ -1,7 +1,7 @@
-import node from "@astrojs/node";
-import { defineConfig } from "astro/config";
+import node from '@astrojs/node';
+import { defineConfig } from 'astro/config';
 
 export default defineConfig({
-  output: "server",
-  adapter: node({ mode: "standalone" }),
+  output: 'server',
+  adapter: node({ mode: 'standalone' }),
 });