+++ packages/opentelemetry/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @astroscope/opentelemetry
 
+## 0.2.0
+
+### Minor Changes
+
+- add @astroscope/excludes for shared exclude patterns; add @astroscope/pino for pino HTTP logging
+- Updated dependencies
+  - @astroscope/excludes@0.1.0
+
 ## 0.1.1
 
 ### Patch Changes